requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
huggingface-hub==0.33.0
idna==3.10
keras==3.10.0
libclang==18.1.1
logger==1.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
